requirements.txt
Flask
render_template
request
os
openai
tiktoken
transformers
googlesearch-python
requests
beautifulsoup4